% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAMC2013to2010.R
\name{convertAMC2013to2010}
\alias{convertAMC2013to2010}
\title{Function to load data from files}
\usage{
convertAMC2013to2010(dataframe, codMun = "IBGE7", nomeMun = "NOME")
}
\arguments{
\item{dataframe}{data with municipal records that will be normalized}

\item{codMun}{name of the column that stores the IBGE code of the municipality}

\item{nomeMun}{column name that stores the municipality name}
}
\value{
dataframe
}
\description{
This function loads existing data into a directory. Data must be formatted as made available by PNUD
}
\details{
In 2013, 5 municipalities were created from the dismemberment of municipal
territories. Longitudinal analyzes or cross-sectional studies that use period variables.
1504752(Moju\\u00ed dos Campos)  -> 1506807(Santar\\u00e9m)
4212650(Pescaria Brava)    -> 4209409(Laguna)
4220000(Balne\\u00e1rio Rinc\\u00e3o)  -> 4207007(I\\u00e7ara)
4314548(Pinto Bandeira)    -> 4302105(Bento Gon\\u00e7alves)
5006275(Para\\u00edso das \\u00c1guas) ->5002951(Chapad\\u00e3o do Sul)```
OBS: The municipality was split using an area of 3 municipalities (\\u00c1gua Clara, Costa Rica and Chapad\\u00e3o do Sul), however it was decided to add it to the most populous municipality;
}
