% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDimensions.R
\name{createDimensions}
\alias{createDimensions}
\title{Create Dimensions}
\usage{
createDimensions(dimensions = c("Ambiental", "Humana"), ..., variables)
}
\arguments{
\item{dimensions}{It must have at least two dimensions, if not informed, the two-dimensional proposition composed of "Environment" and "Human" will be adopted.}

\item{...}{Other optional arguments}

\item{variables}{Database variables that will be included in dimensions}
}
\value{
dataframe
}
\description{
Creates the dataframe with the dimensions and variables that will be computed in the sustainability barometer
}
\details{
The sustainability barometer is a multidimensional metric, therefore, it needs
a minimally two-dimensional structure to be computed. The input dataframe is
organized in a tabular way formed by the dimension, variable and value attributes.
The record (row) of the dataframe is the territories for which the indicators
will be calculated.
}
