% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDictionary.R
\name{dataDictionary}
\alias{dataDictionary}
\title{Data dictionary for the packet data}
\usage{
dataDictionary(text = NULL, ...)
}
\arguments{
\item{text}{search string. Pattern.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The package data is named by acronyms and this package provides the data
dictionary containing the acronym, the indicator name and a more detailed description.
}
