% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a metaexpression}
\usage{
evaluate(x, ...)
}
\arguments{
\item{x}{string that will be evaluated}

\item{...}{Other optional arguments}
}
\description{
In case it is necessary to evaluate a string as a variable
}
\details{
At various times, arguments are expected from functions that are later evaluated
as objects or even functions. This function is a contracted form for the
combination of eval(parse(text="str")).
}
