% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFiles.R
\name{loadDataFromFiles}
\alias{loadDataFromFiles}
\title{Função para carregar os dados a partir dos arquivos}
\usage{
loadDataFromFiles(diretorio = "data")
}
\arguments{
\item{diretorio}{Uma string com o caminho para o diretório desejado.}
}
\description{
Esta função carrega os dados existentes em um diretório.
Os dados precisam estar formatados conforme disponibilizados pela PNUD.
}
