% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naFilter.R
\name{naFilter}
\alias{naFilter}
\title{Filter collumns by NA ratio}
\usage{
naFilter(x, percentCases = 0.5)
}
\arguments{
\item{x}{Dataframe}

\item{percentCases}{threshold}
}
\description{
Select columns that have NA ratio above a threshold. A vector will be
returned with the indices of the columns that exceeded the limit.
}
