% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFiles.R
\name{renewRD}
\alias{renewRD}
\title{Updates the data used in the package}
\usage{
renewRD()
}
\description{
The data used are made available by autonomous data sources such as IBGE,
IPEA, UNDP and others. Whenever there is the possibility of updating these
sources, this function updates the internal Rdata used by the
sustainability barometer.
}
