% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAMC2013to2010.R
\name{convertAMC2013to2010}
\alias{convertAMC2013to2010}
\title{Função para carregar os dados a partir dos arquivos}
\usage{
convertAMC2013to2010(dataframe, codMun = "IBGE7", nomeMun = "NOME")
}
\arguments{
\item{dataframe}{dados com registros municipais que serão normalizados}

\item{codMun}{nome da coluna que armazena o código IBGE do município}

\item{nomeMun}{nome da coluna que armazena o nome do município}
}
\description{
Esta função carrega os dados existentes em um diretório. Os dados precisam estar formatados conforme disponibilizados pela PNUD
}
\details{
No ano de 2013 foram criados 5 municípios a partir do desmembramento de territórios
municipais. Análises longitudinais ou estudos transversais que utilizem variáveis de períodos
Identificando os municípios novos
\code{regAdm[!(unique(regAdm$IBGE7) \%in\% unique(censo$Codmun7)),c("IBGE7","NOME")]}
Pesquisa de onde o município foi desmembrado e localiza o código do IBGE\if{html}{\out{<div class="subset(regAdm, NOME == "Santarém" & substr(regAdm$IBGE7,1,2)=="15", select = c("ANO","IBGE7","NOME"))">}}\preformatted{1504752(Mojuí dos Campos)  -> 1506807(Santarém)
4212650(Pescaria Brava)    -> 4209409(Laguna)
4220000(Balneário Rincão)  -> 4207007(Içara)
4314548(Pinto Bandeira)    -> 4302105(Bento Gonçalves)
5006275(Paraíso das Águas) ->5002951(Chapadão do Sul)```
OBS: O município foi desmembrado utilizando área de 3 municípios (Água Clara,
Costa Rica e Chapadão do Sul) no entanto optou-se por agrega-lo ao município mais populoso;

[!(unique(regAdm$IBGE7) \%in\% unique(censo$Codmun7)),c("IBGE7","NOME")]: R:!(unique(regAdm$IBGE7)\%20\%25in\%25\%20unique(censo$Codmun7)),c(\%22IBGE7\%22,\%22NOME\%22)
}\if{html}{\out{</div>}}
}
