% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFiles.R
\name{loadDataFromFiles}
\alias{loadDataFromFiles}
\title{Function to load raw data from files}
\usage{
loadDataFromFiles(
  ...,
  runConverteAMC = TRUE,
  yearCensus = 2010,
  yearRegAdm = 2013
)
}
\arguments{
\item{...}{optional arguments}

\item{runConverteAMC}{maintains a minimum area of comparison for different territorial configurations of the database}

\item{yearCensus}{year of census data. Includes 2000 and 2010.}

\item{yearRegAdm}{year of administrative records data. Includes 2013 to 2016.}
}
\value{
None
}
\description{
This function loads existing data into a directory.
Data must be formatted as made available by PNUD.
}
\details{
Data source: http://www.atlasbrasil.org.br/acervo/biblioteca
}
